Added two files that connect terminal windows.
These files allow you to type into the client terminal and see it come out on the server terminal.  It appears that at any time, either the server is waiting on the client or the client is waiting on the server.  So for bi-directional data, the two sides will have to take turns.